requirements.txt
flask
pymysql
pyserial
cryptography